app/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Caching>=2.3.0
requests>=2.31.0
google-genai>=1.0.0
gunicorn>=22.0.0
tenacity>=8.2.0
altair>=5.0.0
pytest>=8.0.0
PyYAML>=6.0.0
pycountry>=23.12.11
